gantasala_sai_roopesh-main/requirements.txt
streamlit
PyMuPDF
sentence_transformers
faiss-cpu
numpy
python-dotenv
google-cloud-aiplatform
pdfminer.six
fpdf